Просто переносим рендеринг доски и карточек из `main.js` в `board-controller.js'.
